WIP: split test class
