+++ src/main/resources/com/exclamationlabs/connid/box/Message.properties
@@ -4,7 +4,6 @@
 # Proprietary and confidential
 #
 #
-
 box.connector.display=FIS Charlotte Connector
 basic.group=Basic Configuration Properties
 filePathProperty.display=XML File Path
