+++ src/test/java/com/exclamationlabs/connid/box/GroupGetTests.java
@@ -1,0 +1,85 @@
+/*
+ * Copyright (C) Exclamation Labs 2019. - All Rights Reserved
+ * Unauthorized copying of this file, via any medium is strictly prohibited
+ * Proprietary and confidential
+ *
+ */
+
+package com.exclamationlabs.connid.box;
+
+import com.box.sdk.BoxAPIRequest;
+import com.exclamationlabs.connid.box.testutil.LocalBoxConnector;
+import com.exclamationlabs.connid.box.testutil.MockBoxAPIHelper;
+import org.identityconnectors.framework.api.APIConfiguration;
+import org.identityconnectors.framework.api.ConnectorFacade;
+import org.identityconnectors.framework.api.ConnectorFacadeFactory;
+import org.identityconnectors.framework.common.objects.ConnectorObject;
+import org.identityconnectors.framework.common.objects.Name;
+import org.identityconnectors.framework.common.objects.OperationOptionsBuilder;
+import org.identityconnectors.framework.common.objects.Uid;
+import org.identityconnectors.test.common.TestHelpers;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import java.util.concurrent.atomic.AtomicReference;
+
+import static com.exclamationlabs.connid.box.GroupsHandler.OBJECT_CLASS_GROUP;
+import static com.exclamationlabs.connid.box.testutil.TestUtils.ok;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
+/**
+ * @author Hiroyuki Wada
+ */
+class GroupGetTests {
+
+    ConnectorFacade connector;
+    MockBoxAPIHelper mockAPI;
+
+    @BeforeEach
+    void setup() {
+        connector = newFacade();
+        mockAPI = MockBoxAPIHelper.instance();
+        mockAPI.init();
+    }
+
+    protected ConnectorFacade newFacade() {
+        ConnectorFacadeFactory factory = ConnectorFacadeFactory.getInstance();
+        APIConfiguration impl = TestHelpers.createTestConfiguration(LocalBoxConnector.class, new BoxConfiguration());
+        impl.getResultsHandlerConfiguration().setEnableAttributesToGetSearchResultsHandler(false);
+        impl.getResultsHandlerConfiguration().setEnableNormalizingResultsHandler(false);
+        impl.getResultsHandlerConfiguration().setEnableFilteredResultsHandler(false);
+        return factory.newInstance(impl);
+    }
+
+    @Test
+    void getGroup() {
+        // Given
+        String uid = "11446498";
+        String groupName = "Support";
+
+        AtomicReference<BoxAPIRequest> request = new AtomicReference<>();
+        mockAPI.push(req -> {
+            request.set(req);
+
+            return ok("group-get.json");
+        });
+        mockAPI.push(req -> {
+            return ok("group-member-0.json");
+        });
+
+        // When
+        ConnectorObject result = connector.getObject(OBJECT_CLASS_GROUP,
+                new Uid(uid, new Name(groupName)),
+                new OperationOptionsBuilder().build());
+
+        // Then
+        assertNotNull(request.get());
+        assertEquals("/2.0/groups/" + uid, request.get().getUrl().getPath());
+        assertEquals(OBJECT_CLASS_GROUP, result.getObjectClass());
+        assertEquals(uid, result.getUid().getUidValue());
+        assertEquals(groupName, result.getName().getNameValue());
+        assertNotNull(result.getAttributeByName("description"));
+        assertEquals("Support Group - as imported from Active Directory", result.getAttributeByName("description").getValue().get(0));
+    }
+}
