+++ src/test/java/com/exclamationlabs/connid/box/GroupUpdateTests.java
@@ -1,0 +1,139 @@
+/*
+ * Copyright (C) Exclamation Labs 2019. - All Rights Reserved
+ * Unauthorized copying of this file, via any medium is strictly prohibited
+ * Proprietary and confidential
+ *
+ */
+
+package com.exclamationlabs.connid.box;
+
+import com.box.sdk.BoxAPIRequest;
+import com.exclamationlabs.connid.box.testutil.LocalBoxConnector;
+import com.exclamationlabs.connid.box.testutil.MockBoxAPIHelper;
+import org.identityconnectors.framework.api.APIConfiguration;
+import org.identityconnectors.framework.api.ConnectorFacade;
+import org.identityconnectors.framework.api.ConnectorFacadeFactory;
+import org.identityconnectors.framework.common.exceptions.RetryableException;
+import org.identityconnectors.framework.common.exceptions.UnknownUidException;
+import org.identityconnectors.framework.common.objects.*;
+import org.identityconnectors.test.common.TestHelpers;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import java.util.HashSet;
+import java.util.Set;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicReference;
+
+import static com.exclamationlabs.connid.box.GroupsHandler.OBJECT_CLASS_GROUP;
+import static com.exclamationlabs.connid.box.testutil.TestUtils.*;
+import static org.junit.jupiter.api.Assertions.*;
+
+/**
+ * @author Hiroyuki Wada
+ */
+class GroupUpdateTests {
+
+    ConnectorFacade connector;
+    MockBoxAPIHelper mockAPI;
+
+    @BeforeEach
+    void setup() {
+        connector = newFacade();
+        mockAPI = MockBoxAPIHelper.instance();
+        mockAPI.init();
+    }
+
+    protected ConnectorFacade newFacade() {
+        ConnectorFacadeFactory factory = ConnectorFacadeFactory.getInstance();
+        APIConfiguration impl = TestHelpers.createTestConfiguration(LocalBoxConnector.class, new BoxConfiguration());
+        impl.getResultsHandlerConfiguration().setEnableAttributesToGetSearchResultsHandler(false);
+        impl.getResultsHandlerConfiguration().setEnableNormalizingResultsHandler(false);
+        impl.getResultsHandlerConfiguration().setEnableFilteredResultsHandler(false);
+        return factory.newInstance(impl);
+    }
+
+    @Test
+    void updateGroup() {
+        // Given
+        String groupName = "Support";
+
+        Set<AttributeDelta> modifications = new HashSet<>();
+        modifications.add(AttributeDeltaBuilder.build("description", "Support Group"));
+
+        AtomicReference<BoxAPIRequest> request = new AtomicReference<>();
+        mockAPI.push(req -> {
+            request.set(req);
+
+            return ok("group-update.json");
+        });
+
+        // When
+        Set<AttributeDelta> sideEffects = connector.updateDelta(OBJECT_CLASS_GROUP,
+                new Uid("11446498", new Name(groupName)),
+                modifications, new OperationOptionsBuilder().build());
+
+        // Then
+        assertNotNull(request.get());
+        assertEquals("Support Group", getJsonAttr(request.get(), "description"));
+        assertNull(sideEffects);
+    }
+
+    @Test
+    void updateGroup_notFound() {
+        // Given
+        String groupName = "Support";
+
+        Set<AttributeDelta> modifications = new HashSet<>();
+        modifications.add(AttributeDeltaBuilder.build("description", "Support Group"));
+
+        mockAPI.push(req -> {
+            throw notFound();
+        });
+
+        // When
+        UnknownUidException e = assertThrows(UnknownUidException.class, () -> {
+            Set<AttributeDelta> sideEffects = connector.updateDelta(OBJECT_CLASS_GROUP,
+                    new Uid("11446498", new Name(groupName)),
+                    modifications, new OperationOptionsBuilder().build());
+        });
+
+        // Then
+        assertNotNull(e);
+    }
+
+    @Test
+    void updateGroup_otherError() {
+        // Given
+        String groupName = "Support";
+
+        Set<AttributeDelta> modifications = new HashSet<>();
+        modifications.add(AttributeDeltaBuilder.build("description", "Support Group"));
+
+        // Set retry count of the Box SDK
+        mockAPI.setMaxRequestAttempts(2);
+
+        AtomicInteger count = new AtomicInteger();
+        mockAPI.push(req -> {
+            count.incrementAndGet();
+
+            throw internalServerError();
+        });
+        mockAPI.push(req -> {
+            count.incrementAndGet();
+
+            throw internalServerError();
+        });
+
+        // When
+        RetryableException e = assertThrows(RetryableException.class, () -> {
+            Set<AttributeDelta> sideEffects = connector.updateDelta(OBJECT_CLASS_GROUP,
+                    new Uid("11446498", new Name(groupName)),
+                    modifications, new OperationOptionsBuilder().build());
+        });
+
+        // Then
+        assertNotNull(e);
+        assertEquals(2, count.get());
+    }
+}
