+++ src/test/java/com/exclamationlabs/connid/box/UserCreateTests.java
@@ -1,0 +1,107 @@
+/*
+ * Copyright (C) Exclamation Labs 2019. - All Rights Reserved
+ * Unauthorized copying of this file, via any medium is strictly prohibited
+ * Proprietary and confidential
+ *
+ */
+
+package com.exclamationlabs.connid.box;
+
+import com.box.sdk.BoxAPIRequest;
+import com.exclamationlabs.connid.box.testutil.LocalBoxConnector;
+import com.exclamationlabs.connid.box.testutil.MockBoxAPIHelper;
+import org.identityconnectors.framework.api.APIConfiguration;
+import org.identityconnectors.framework.api.ConnectorFacade;
+import org.identityconnectors.framework.api.ConnectorFacadeFactory;
+import org.identityconnectors.framework.common.exceptions.AlreadyExistsException;
+import org.identityconnectors.framework.common.objects.*;
+import org.identityconnectors.test.common.TestHelpers;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import java.util.HashSet;
+import java.util.Set;
+import java.util.concurrent.atomic.AtomicReference;
+
+import static com.exclamationlabs.connid.box.UsersHandler.OBJECT_CLASS_USER;
+import static com.exclamationlabs.connid.box.testutil.TestUtils.*;
+import static org.junit.jupiter.api.Assertions.*;
+
+/**
+ * @author Hiroyuki Wada
+ */
+class UserCreateTests {
+
+    ConnectorFacade connector;
+    MockBoxAPIHelper mockAPI;
+
+    @BeforeEach
+    void setup() {
+        connector = newFacade();
+        mockAPI = MockBoxAPIHelper.instance();
+        mockAPI.init();
+    }
+
+    protected ConnectorFacade newFacade() {
+        ConnectorFacadeFactory factory = ConnectorFacadeFactory.getInstance();
+        APIConfiguration impl = TestHelpers.createTestConfiguration(LocalBoxConnector.class, new BoxConfiguration());
+        impl.getResultsHandlerConfiguration().setEnableAttributesToGetSearchResultsHandler(false);
+        impl.getResultsHandlerConfiguration().setEnableNormalizingResultsHandler(false);
+        impl.getResultsHandlerConfiguration().setEnableFilteredResultsHandler(false);
+        return factory.newInstance(impl);
+    }
+
+    @Test
+    void createUser() {
+        // Given
+        String login = "ceo@example.com";
+        String name = "Aaron Levie";
+
+        Set<Attribute> attributes = new HashSet<>();
+        attributes.add(new Name(login));
+        attributes.add(AttributeBuilder.build("name", name));
+
+        AtomicReference<BoxAPIRequest> request = new AtomicReference<>();
+        mockAPI.push(req -> {
+            request.set(req);
+
+            return created("user-create.json");
+        });
+
+        // When
+        Uid uid = connector.create(OBJECT_CLASS_USER, attributes, new OperationOptionsBuilder().build());
+
+        // Then
+        assertNotNull(request.get());
+        assertEquals(login, getJsonAttr(request.get(), "login"));
+        assertEquals(name, getJsonAttr(request.get(), "name"));
+        assertEquals("11446498", uid.getUidValue());
+        assertEquals(login, uid.getNameHintValue());
+    }
+
+    @Test
+    void createUser_alreadyExists() {
+        // Given
+        String login = "ceo@example.com";
+        String name = "Aaron Levie";
+
+        Set<Attribute> attributes = new HashSet<>();
+        attributes.add(new Name(login));
+        attributes.add(AttributeBuilder.build("name", name));
+
+        AtomicReference<BoxAPIRequest> request = new AtomicReference<>();
+        mockAPI.push(req -> {
+            request.set(req);
+
+            throw conflict();
+        });
+
+        // When
+        AlreadyExistsException e = assertThrows(AlreadyExistsException.class, () -> {
+            Uid uid = connector.create(OBJECT_CLASS_USER, attributes, new OperationOptionsBuilder().build());
+        });
+
+        // Then
+        assertNotNull(e);
+    }
+}
